+++ editor/graph/voxel_graph_editor.cpp
@@ -32,6 +32,7 @@
 #include "../../util/math/conv.h"
 #include "../../util/profiling.h"
 #include "../../util/string/format.h"
+#include "graph_editor_adapter.h"
 #include "voxel_graph_editor_node.h"
 #include "voxel_graph_editor_node_preview.h"
 #include "voxel_graph_editor_shader_dialog.h"
@@ -785,13 +786,13 @@ void VoxelGraphEditor::_on_menu_id_pressed(int id) {
 
 #ifdef VOXEL_ENABLE_GPU
 		case MENU_GENERATE_SHADER: {
-			ERR_FAIL_COND(_generator.is_null());
-			VoxelGenerator::ShaderSourceData sd;
-			if (!_generator->get_shader_source(sd)) {
+			ERR_FAIL_COND(_graph.is_null());
+			pg::VoxelGraphFunction::ShaderResult shader_res = _graph->get_shader_source();
+			if (!shader_res.compilation.success) {
 				return;
 			}
 			// TODO Include uniforms in that version?
-			_shader_dialog->set_shader_code(sd.glsl);
+			_shader_dialog->set_shader_code(to_godot(shader_res.code_utf8));
 			_shader_dialog->popup_centered();
 		} break;
 #endif
@@ -1003,9 +1004,16 @@ void VoxelGraphEditor::update_previews(bool with_live_update) {
 		_compile_result_label->hide();
 	}
 
-	if (_generator.is_null() || !_generator->is_good()) {
-		return;
+	if (_generator.is_valid()) {
+		if (!_generator->is_good()) {
+			return;
+		}
+	} else {
+		if (!_graph->is_compiled()) {
+			return;
+		}
 	}
+
 	// We assume no other thread will try to modify the graph and compile something not good
 
 	// TODO Make slice previews work with arbitrary functions, when possible
@@ -1043,15 +1051,13 @@ void VoxelGraphEditor::update_previews(bool with_live_update) {
 
 void VoxelGraphEditor::update_range_analysis_previews() {
 	ZN_PRINT_VERBOSE("Updating range analysis previews");
-	ERR_FAIL_COND(_generator.is_null());
-	ERR_FAIL_COND(!_generator->is_good());
+	GraphEditorAdapter adapter(_generator, _graph);
+	ERR_FAIL_COND(!adapter.is_good());
 
 	const AABB aabb = _range_analysis_dialog->get_aabb();
-	_generator->debug_analyze_range(
-			math::floor_to_int(aabb.position), math::floor_to_int(aabb.position + aabb.size), true
-	);
+	adapter.debug_analyze_range(math::floor_to_int(aabb.position), math::floor_to_int(aabb.position + aabb.size));
 
-	const pg::Runtime::State &state = _generator->get_last_state_from_current_thread();
+	const pg::Runtime::State &state = adapter.get_last_state_from_current_thread();
 
 	const Color greyed_out_color(1, 1, 1, 0.5);
 
@@ -1069,7 +1075,7 @@ void VoxelGraphEditor::update_range_analysis_previews() {
 		// TODO Would be nice if GraphEdit's minimap would take such coloring into account...
 		node_view->set_modulate(greyed_out_color);
 
-		node_view->update_range_analysis_tooltips(**_generator, state);
+		node_view->update_range_analysis_tooltips(adapter, state);
 	}
 
 	// Highlight only nodes that will actually run.
@@ -1116,7 +1122,8 @@ void VoxelGraphEditor::update_range_analysis_gizmo() {
 void VoxelGraphEditor::update_slice_previews() {
 	// TODO Use a thread?
 	ZN_PRINT_VERBOSE("Updating slice previews");
-	ERR_FAIL_COND(!_generator->is_good());
+
+	GraphEditorAdapter adapter(_generator, _graph);
 
 	struct PreviewInfo {
 		VoxelGraphEditorNodePreview *control;
@@ -1142,7 +1149,7 @@ void VoxelGraphEditor::update_slice_previews() {
 		}
 		PreviewInfo info;
 		info.control = node->get_preview();
-		if (!_generator->try_get_output_port_address(src, info.address)) {
+		if (!adapter.try_get_output_port_address(src, info.address)) {
 			// Not part of the compiled result
 			continue;
 		}
@@ -1154,6 +1161,7 @@ void VoxelGraphEditor::update_slice_previews() {
 	{
 		const int preview_size_x = VoxelGraphEditorNodePreview::RESOLUTION;
 		const int preview_size_y = VoxelGraphEditorNodePreview::RESOLUTION;
+		// TODO This might be too big for buffer size?
 		const int buffer_size = preview_size_x * preview_size_y;
 		StdVector<float> x_vec;
 		StdVector<float> y_vec;
@@ -1191,10 +1199,10 @@ void VoxelGraphEditor::update_slice_previews() {
 			z_coords = to_span(y_vec);
 		}
 
-		_generator->generate_set(x_coords, y_coords, z_coords);
+		adapter.generate_set(x_coords, y_coords, z_coords);
 	}
 
-	const pg::Runtime::State &last_state = VoxelGeneratorGraph::get_last_state_from_current_thread();
+	const pg::Runtime::State &last_state = adapter.get_last_state_from_current_thread();
 
 	// Update previews
 	for (size_t preview_index = 0; preview_index < previews.size(); ++preview_index) {
@@ -1324,8 +1332,19 @@ void VoxelGraphEditor::hide_profiling_ratios() {
 
 void VoxelGraphEditor::update_buttons_availability() {
 	// Some features are only available with a generator (for now)
-	_debug_menu_button->set_disabled(_generator.is_null());
-	_graph_menu_button->set_disabled(_generator.is_null());
+	// _debug_menu_button->set_disabled(_generator.is_null());
+	// _graph_menu_button->set_disabled(_generator.is_null());
+
+	// TODO Implement profiling on any graph
+	PopupMenu &menu = *_debug_menu_button->get_popup();
+	{
+		const int index = menu.get_item_index(MENU_PROFILE);
+		menu.set_item_disabled(index, _generator.is_null());
+	}
+	{
+		const int index = menu.get_item_index(MENU_LIVE_UPDATE);
+		menu.set_item_disabled(index, _generator.is_null());
+	}
 }
 
 void VoxelGraphEditor::_on_range_analysis_toggled(bool enabled) {